requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --index-url=https://pypi.python.org/simple/ --no-emit-trusted-host --no-index requirements.in
#
asgiref==3.3.1
    # via django
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
django==3.1.4
    # via
    #   -r requirements.in
    #   djangorestframework
djangorestframework==3.12.2
    # via -r requirements.in
idna==2.10
    # via requests
postmarker==0.17.0
    # via -r requirements.in
pytz==2020.5
    # via django
requests==2.25.1
    # via postmarker
sqlparse==0.4.1
    # via django
urllib3==1.26.2
    # via requests
